+++ domain/src/main/java/be/kuleuven/cs/flexsim/domain/energy/dso/r3dp/FlexibilityProvider.java
@@ -23,7 +23,7 @@ public interface FlexibilityProvider {
     /**
      * @return The amaount of monetary compensation for activating flexibility.
      */
-    long getMonetaryCompensationValue();
+    double getMonetaryCompensationValue();
 
     /**
      * Register an activation event for this Flexibility provider.
